+++ test/unit/__mocks__/http-proxy.js
@@ -1,9 +1,8 @@
 const httpProxy = jest.genMockFromModule("http-proxy");
 httpProxy.createProxyServer = jest.fn(function() {
 	return {
-		web: jest.fn(function(req, res, next) {
-			next();
-		})
+		web: jest.fn(),
+		on: jest.fn()
 	};
 });
 module.exports = httpProxy;